apps/users_service/requirements/base.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dictdiffer==0.9.0
exceptiongroup==1.1.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.0
pluggy==1.0.0
pydantic==1.10.4
PyJWT==2.6.0
pyotp==2.8.0
pypika-tortoise==0.1.6
pytest==7.2.1
python-multipart==0.0.5
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
